refactor(site): page profile reward
